Acton LMDB module
This adds an lmdb module for Acton which interacts with LMDB. There is a
basic non-transactional interface (implied transactions) for basic
interaction with put/get/del. The more advanced interface spawns
explicit read / write transactions, which are represented by individual
actors. Since LMDB supports multiple read transactions, we allows
multiple ReadTransaction actors while WriteTransaction is limited to 1
and thus we keep a lock on that which we keep track of. This is a new
style for wrapping up a C library in, utilizing multiple actors (we
normally only have a single actor). I'm also using separate standalone
private functions that take a pointer as argument for the actual C
wrapping - I've had this idea for quite some time but this is the first
time I'm really using it - feels pretty good. Using int for the pointer
is a waste but int should be i64 soon so... well, I guess we should have
a ptr type for this anyway.

The cursor interface layers on top of transactions and allows iterating
over keys.

There are test cases for most of the functionality offered.